moved locationVM to it's own class
