+++ BabyBrunch/Views/MapViewHelpers/UIKitMapView.swift
@@ -16,7 +16,7 @@ struct UIKitMapView : UIViewRepresentable {
     @Binding var mapViewRef: MKMapView?
     @Binding var selectedVenue : MKMapItem?
     
-    @StateObject private var vm = ContentViewModel()
+          
     
     
     func makeCoordinator() -> Coordinator {
@@ -78,40 +78,4 @@ struct UIKitMapView : UIViewRepresentable {
     
 }
 
-final class ContentViewModel: NSObject, ObservableObject, CLLocationManagerDelegate {
-    var locationManager : CLLocationManager?
-    
-    func checkIfLocationServicesEnabled() {
-        if CLLocationManager.locationServicesEnabled() {
-            locationManager = CLLocationManager()
-            //    locationManager?.desiredAccuracy = kCLLocationAccuracyBest // add if you need it later
-                // checkLocationAuthorization( ) // redundant
-            locationManager!.delegate = self
-            
-        } else {
-            print( "Location services are not enabled." )
-        }
-        
-    }
-    func checkLocationAuthorization() {
-        guard let locationManager = locationManager else { return }
-      
-        switch locationManager.authorizationStatus {
-            
-        case .notDetermined:
-            locationManager.requestWhenInUseAuthorization( )
-        case .restricted:
-            print("Restricted location")
-        case .denied:
-            print("You denied location")
-        case .authorizedAlways, .authorizedWhenInUse:
-            break
-        @unknown default:
-        break
-        }
-    }
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        checkLocationAuthorization()
-    }
-}
